countryAPI-master/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
djangorestframework==3.14.0
idna==3.4
pytz==2022.6
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.13
wikipedia==1.4.0
